Change nav header to flex
I was debating between using flex and grid for the header. It's still not clear to me what the major use cases are for either. All I know is that the flexbox design I have now looks much better than the grid I had attempted to implement.

Still have to make a lot of this responsive/mobile-friendly.